# main/src/main/resources/openapi/openapi.yaml
...
openapi: 3.0.3
info:
  title: Policy Administration Documentation
  description: Policy Administration is responsible for the deployment life cycle of policies as well as
    interworking with the mechanisms required to orchestrate the nodes and containers on which
    policies run. It is also responsible for the administration of policies at run time;
    ensuring that policies are available to users, that policies are executing correctly,
    and that the state and status of policies is monitored
  termsOfService: urn:tos
  contact:
    name: ONAP Support
    url: https://lists.onap.org/g/onap-discuss
    email: onap-discuss@lists.onap.org
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0
  version: '1.0'
...
servers:
  - url: https://{server}
    variables:
      server:
        default: policy/pap/v1
        description: This value is assigned by the service provider
...
paths:
...
  /pdps/groups/{name}:
    put:
      tags:
        - PdpGroupStateChangeControllerV1
      summary: Change state of a PDP Group
      description: Changes state of PDP Group, returning optional error details
      operationId: changeGroupState
      parameters:
        - name: X-ONAP-RequestID
          in: header
          description: RequestID for http transaction
          required: false
          schema:
            type: string
            format: uuid
        - name: name
          in: path
          description: PDP Group Name on which to change the state
          required: true
          style: simple
          schema:
            type: string
        - name: state
          in: query
          description: PDP Group State, an instance of
            [PdpState](https://github.com/onap/policy-models/blob/master/models-pdp/src/main/java/org/onap/policy/models/pdp/enums/PdpState.java)
          required: true
          style: form
          schema:
            $ref: '#/components/schemas/PdpState'
      responses:
        200:
          description: OK, the result of the state shange operation in an instance of
            [PdpGroupStateChangeResponse](https://github.com/onap/policy-models/blob/master/models-pap/src/main/java/org/onap/policy/models/pap/concepts/PdpGroupStateChangeResponse.java)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PdpGroupStateChangeResponse'
            application/yaml:
              schema:
                $ref: '#/components/schemas/PdpGroupStateChangeResponse'
        201:
          description: Created, the result of the state shange operation in an instance of
            [PdpGroupStateChangeResponse](https://github.com/onap/policy-models/blob/master/models-pap/src/main/java/org/onap/policy/models/pap/concepts/PdpGroupStateChangeResponse.java)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PdpGroupStateChangeResponse'
            application/yaml:
              schema:
                $ref: '#/components/schemas/PdpGroupStateChangeResponse'
        401:
          description: Authentication Error
          headers:
            X-LatestVersion:
              description: Used only to communicate an API's latest version
              required: true
              schema:
                type: string
            X-PatchVersion:
              description: >-
                Used only to communicate a PATCH version in a response for
                troubleshooting purposes only, and will not be provided by the
                client on request
              required: true
              schema:
                type: string
            X-MinorVersion:
              description: >-
                Used to request or communicate a MINOR version back from the
                client to the server, and from the server back to the client
              required: true
              schema:
                type: string
            X-ONAP-RequestID:
              description: Used to track REST transactions for logging purpose
              required: true
              schema:
                type: string
                format: uuid
        403:
          description: Authorization Error
          headers:
            X-LatestVersion:
              description: Used only to communicate an API's latest version
              required: true
              schema:
                type: string
            X-PatchVersion:
              description: >-
                Used only to communicate a PATCH version in a response for
                troubleshooting purposes only, and will not be provided by the
                client on request
              required: true
              schema:
                type: string
            X-MinorVersion:
              description: >-
                Used to request or communicate a MINOR version back from the
                client to the server, and from the server back to the client
              required: true
              schema:
                type: string
            X-ONAP-RequestID:
              description: Used to track REST transactions for logging purpose
              required: true
              schema:
                type: string
                format: uuid
        404:
          description: Not Found
        500:
          description: Internal Server Error
          headers:
            X-LatestVersion:
              description: Used only to communicate an API's latest version
              required: true
              schema:
                type: string
            X-PatchVersion:
              description: >-
                Used only to communicate a PATCH version in a response for
                troubleshooting purposes only, and will not be provided by the
                client on request
              required: true
              schema:
                type: string
            X-MinorVersion:
              description: >-
                Used to request or communicate a MINOR version back from the
                client to the server, and from the server back to the client
              required: true
              schema:
                type: string
            X-ONAP-RequestID:
              description: Used to track REST transactions for logging purpose
              required: true
              schema:
                type: string
                format: uuid
      security:
        - basicAuth: []
      x-interface info:
        api-version: 1.0.0
        last-mod-release: Dublin
    delete:
      tags:
        - PdpGroupDeleteControllerV1
      summary: Delete PDP Group
      description: Deletes a PDP Group, returning optional error details
      operationId: deleteGroup
      parameters:
        - name: X-ONAP-RequestID
          in: header
          description: RequestID for http transaction
          required: false
          schema:
            type: string
            format: uuid
        - name: name
          in: path
          description: PDP Group Name
          required: true
          style: simple
          schema:
            type: string
      responses:
        200:
          description: OK, the result of the delete operation in an instance of
            [PdpGroupDeleteResponse](https://github.com/onap/policy-models/blob/master/models-pap/src/main/java/org/onap/policy/models/pap/concepts/PdpGroupDeleteResponse.java)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PdpGroupDeleteResponse'
            application/yaml:
              schema:
                $ref: '#/components/schemas/PdpGroupDeleteResponse'
        204:
          description: No Content
        401:
          description: Authentication Error
          headers:
            X-LatestVersion:
              description: Used only to communicate an API's latest version
              required: true
              schema:
                type: string
            X-PatchVersion:
              description: >-
                Used only to communicate a PATCH version in a response for
                troubleshooting purposes only, and will not be provided by the
                client on request
              required: true
              schema:
                type: string
            X-MinorVersion:
              description: >-
                Used to request or communicate a MINOR version back from the
                client to the server, and from the server back to the client
              required: true
              schema:
                type: string
            X-ONAP-RequestID:
              description: Used to track REST transactions for logging purpose
              required: true
              schema:
                type: string
                format: uuid
        403:
          description: Authorization Error
          headers:
            X-LatestVersion:
              description: Used only to communicate an API's latest version
              required: true
              schema:
                type: string
            X-PatchVersion:
              description: >-
                Used only to communicate a PATCH version in a response for
                troubleshooting purposes only, and will not be provided by the
                client on request
              required: true
              schema:
                type: string
            X-MinorVersion:
              description: >-
                Used to request or communicate a MINOR version back from the
                client to the server, and from the server back to the client
              required: true
              schema:
                type: string
            X-ONAP-RequestID:
              description: Used to track REST transactions for logging purpose
              required: true
              schema:
                type: string
                format: uuid
        500:
          description: Internal Server Error
          headers:
            X-LatestVersion:
              description: Used only to communicate an API's latest version
              required: true
              schema:
                type: string
            X-PatchVersion:
              description: >-
                Used only to communicate a PATCH version in a response for
                troubleshooting purposes only, and will not be provided by the
                client on request
              required: true
              schema:
                type: string
            X-MinorVersion:
              description: >-
                Used to request or communicate a MINOR version back from the
                client to the server, and from the server back to the client
              required: true
              schema:
                type: string
            X-ONAP-RequestID:
              description: Used to track REST transactions for logging purpose
              required: true
              schema:
                type: string
                format: uuid
      security:
        - basicAuth: []
      x-interface info:
        api-version: 1.0.0
        last-mod-release: Dublin
  /pdps/policies/{name}:
    delete:
      tags:
        - PdpGroupDeleteControllerV1
      summary: Undeploy a PDP Policy from PDPs
      description: >-
        Undeploys the latest version of a policy from the PDPs, returning
        optional error details
      operationId: deletePolicy
      parameters:
        - name: X-ONAP-RequestID
          in: header
          description: RequestID for http transaction
          required: false
          schema:
            type: string
            format: uuid
        - name: name
          in: path
          description: PDP Policy Name
          required: true
          style: simple
          schema:
            type: string
      responses:
        200:
          description: OK, the result of the undeploy operation in an instance of
            [PdpGroupDeployResponse](https://github.com/onap/policy-models/blob/master/models-pap/src/main/java/org/onap/policy/models/pap/concepts/PdpGroupDeployResponse.java)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PdpGroupDeployResponse'
            application/yaml:
              schema:
                $ref: '#/components/schemas/PdpGroupDeployResponse'
        204:
          description: No Content
        401:
          description: Authentication Error
          headers:
            X-LatestVersion:
              description: Used only to communicate an API's latest version
              required: true
              schema:
                type: string
            X-PatchVersion:
              description: >-
                Used only to communicate a PATCH version in a response for
                troubleshooting purposes only, and will not be provided by the
                client on request
              required: true
              schema:
                type: string
            X-MinorVersion:
              description: >-
                Used to request or communicate a MINOR version back from the
                client to the server, and from the server back to the client
              required: true
              schema:
                type: string
            X-ONAP-RequestID:
              description: Used to track REST transactions for logging purpose
              required: true
              schema:
                type: string
                format: uuid
        403:
          description: Authorization Error
          headers:
            X-LatestVersion:
              description: Used only to communicate an API's latest version
              required: true
              schema:
                type: string
            X-PatchVersion:
              description: >-
                Used only to communicate a PATCH version in a response for
                troubleshooting purposes only, and will not be provided by the
                client on request
              required: true
              schema:
                type: string
            X-MinorVersion:
              description: >-
                Used to request or communicate a MINOR version back from the
                client to the server, and from the server back to the client
              required: true
              schema:
                type: string
            X-ONAP-RequestID:
              description: Used to track REST transactions for logging purpose
              required: true
              schema:
                type: string
                format: uuid
        500:
          description: Internal Server Error
          headers:
            X-LatestVersion:
              description: Used only to communicate an API's latest version
              required: true
              schema:
                type: string
            X-PatchVersion:
              description: >-
                Used only to communicate a PATCH version in a response for
                troubleshooting purposes only, and will not be provided by the
                client on request
              required: true
              schema:
                type: string
            X-MinorVersion:
              description: >-
                Used to request or communicate a MINOR version back from the
                client to the server, and from the server back to the client
              required: true
              schema:
                type: string
            X-ONAP-RequestID:
              description: Used to track REST transactions for logging purpose
              required: true
              schema:
                type: string
                format: uuid
      security:
        - basicAuth: []
      x-interface info:
        api-version: 1.0.0
        last-mod-release: Dublin
  /pdps/policies/{name}/versions/{version}:
    delete:
      tags:
        - PdpGroupDeleteControllerV1
      summary: Undeploy version of a PDP Policy from PDPs
      description: >-
        Undeploys a specific version of a policy from the PDPs, returning
        optional error details
      operationId: deletePolicyVersion
      parameters:
        - name: X-ONAP-RequestID
          in: header
          description: RequestID for http transaction
          required: false
          schema:
            type: string
            format: uuid
        - name: name
          in: path
          description: PDP Policy Name
          required: true
          style: simple
          schema:
            type: string
        - name: version
          in: path
          description: PDP Policy Version
          required: true
          style: simple
          schema:
            type: string
      responses:
        200:
          description: OK, the result of the undeploy operation in an instance of
            [PdpGroupDeployResponse](https://github.com/onap/policy-models/blob/master/models-pap/src/main/java/org/onap/policy/models/pap/concepts/PdpGroupDeployResponse.java)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PdpGroupDeployResponse'
            application/yaml:
              schema:
                $ref: '#/components/schemas/PdpGroupDeployResponse'
        204:
          description: No Content
        401:
          description: Authentication Error
          headers:
            X-LatestVersion:
              description: Used only to communicate an API's latest version
              required: true
              schema:
                type: string
            X-PatchVersion:
              description: >-
                Used only to communicate a PATCH version in a response for
                troubleshooting purposes only, and will not be provided by the
                client on request
              required: true
              schema:
                type: string
            X-MinorVersion:
              description: >-
                Used to request or communicate a MINOR version back from the
                client to the server, and from the server back to the client
              required: true
              schema:
                type: string
            X-ONAP-RequestID:
              description: Used to track REST transactions for logging purpose
              required: true
              schema:
                type: string
                format: uuid
        403:
          description: Authorization Error
          headers:
            X-LatestVersion:
              description: Used only to communicate an API's latest version
              required: true
              schema:
                type: string
            X-PatchVersion:
              description: >-
                Used only to communicate a PATCH version in a response for
                troubleshooting purposes only, and will not be provided by the
                client on request
              required: true
              schema:
                type: string
            X-MinorVersion:
              description: >-
                Used to request or communicate a MINOR version back from the
                client to the server, and from the server back to the client
              required: true
              schema:
                type: string
            X-ONAP-RequestID:
              description: Used to track REST transactions for logging purpose
              required: true
              schema:
                type: string
                format: uuid
        500:
          description: Internal Server Error
          headers:
            X-LatestVersion:
              description: Used only to communicate an API's latest version
              required: true
              schema:
                type: string
            X-PatchVersion:
              description: >-
                Used only to communicate a PATCH version in a response for
                troubleshooting purposes only, and will not be provided by the
                client on request
              required: true
              schema:
                type: string
            X-MinorVersion:
              description: >-
                Used to request or communicate a MINOR version back from the
                client to the server, and from the server back to the client
              required: true
              schema:
                type: string
            X-ONAP-RequestID:
              description: Used to track REST transactions for logging purpose
              required: true
              schema:
                type: string
                format: uuid
      security:
        - basicAuth: []
      x-interface info:
        api-version: 1.0.0
        last-mod-release: Dublin
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
...
    PdpState:
      title: PdpState
      type: object
...
    PdpGroupStateChangeResponse:
      title: PdpGroupStateChangeResponse
      type: object
    PdpGroupDeleteResponse:
      title: PdpGroupDeleteResponse
      type: object
    PdpGroupDeployResponse:
      title: PdpGroupDeployResponse
      type: object
...